load sk and trading client separately
