+++ src/components/header/elements/settings-block/components/one-click-trading/components/OneClickTradingModal.tsx
@@ -223,9 +223,6 @@ export const OneClickTradingModal = ({ isOpen, onClose }: OneClickTradingModalPr
         return;
       }
     }
-    if (walletClient) {
-      setTradingClient(walletClient);
-    }
   }, [address, walletClient, storageKey, activatedOneClickTrading, setTradingClient]);
 
   return (
