+++ src/pages/trader-page/TraderPage.tsx
@@ -42,6 +42,7 @@ import { PerpetualInfoFetcher } from './components/PerpetualInfoFetcher';
 import { PoolSubscription } from './components/PoolSubscription';
 import { CandlesWebSocketListener } from './components/candles-webSocket-listener/CandlesWebSocketListener';
 import { TableDataFetcher } from './components/table-data-refetcher/TableDataFetcher';
+import { SDKLoader } from './components/sdk-loader/SDKLoader';
 
 import styles from './TraderPage.module.scss';
 
@@ -336,6 +337,7 @@ export const TraderPage = () => {
         </MaintenanceWrapper>
       </Box>
 
+      <SDKLoader />
       <UsdcSwapModal isOpen={dialogOpen} onClose={closeDialog} />
       <TableDataFetcher />
       <PerpetualInfoFetcher />
